+++ static/arguments.yml
@@ -462,6 +462,21 @@
     type: Hash
     default: N/A
     description: "[System arguments](/system-arguments)"
+- component: ButtonGroup
+  source: https://github.com/primer/view_components/tree/main/app/components/primer/beta/button_group.rb
+  parameters:
+  - name: variant
+    type: Symbol
+    default: "`nil`"
+    description: DEPRECATED. One of `:medium` and `:small`.
+  - name: size
+    type: Symbol
+    default: "`:medium`"
+    description: One of `:medium` and `:small`.
+  - name: system_arguments
+    type: Hash
+    default: N/A
+    description: "[System arguments](/system-arguments)"
 - component: Flash
   source: https://github.com/primer/view_components/tree/main/app/components/primer/beta/flash.rb
   parameters:
@@ -550,21 +565,6 @@
     type: Boolean
     default: "`false`"
     description: Whether or not to render a dropdown caret.
-  - name: system_arguments
-    type: Hash
-    default: N/A
-    description: "[System arguments](/system-arguments)"
-- component: ButtonGroup
-  source: https://github.com/primer/view_components/tree/main/app/components/primer/button_group.rb
-  parameters:
-  - name: variant
-    type: Symbol
-    default: "`nil`"
-    description: DEPRECATED. One of `:medium` and `:small`.
-  - name: size
-    type: Symbol
-    default: "`:medium`"
-    description: One of `:medium` and `:small`.
   - name: system_arguments
     type: Hash
     default: N/A
