+++ static/statuses.json
@@ -22,6 +22,7 @@
   "Primer::Beta::BorderBox::Header": "alpha",
   "Primer::Beta::Breadcrumbs": "beta",
   "Primer::Beta::Breadcrumbs::Item": "alpha",
+  "Primer::Beta::ButtonGroup": "beta",
   "Primer::Beta::Flash": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
@@ -30,7 +31,7 @@
   "Primer::BorderBoxComponent": "deprecated",
   "Primer::BoxComponent": "stable",
   "Primer::ButtonComponent": "beta",
-  "Primer::ButtonGroup": "beta",
+  "Primer::ButtonGroup": "deprecated",
   "Primer::ClipboardCopy": "beta",
   "Primer::CloseButton": "beta",
   "Primer::ConditionalWrapper": "alpha",
